% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update iNZightVIT}
\usage{
update(os = c("windows", "macos", "linux"))
}
\arguments{
\item{os}{one of 'windows', 'macos', or 'linux'}
}
\value{
logical whether or not update was successful
}
\description{
Update iNZightVIT
}
\author{
Tom Elliott
}
